fix(validator): Correct parameter order in `positiveLike()` method for consistent type handling of numeric values.

+++ src/Base/BaseValidator.php
@@ -205,7 +205,7 @@ abstract class BaseValidator
      * // `true`
      * ```
      */
-    public static function positiveLike(int|float|string|Stringable $value, float|null $max = null): bool
+    public static function positiveLike(float|int|string|Stringable $value, float|null $max = null): bool
     {
         if (is_int($value) || is_float($value)) {
             return $value > 0.0 && ($max === null || $value <= $max);
